+++ Frends.SFTP.MoveFile/Frends.SFTP.MoveFile.Tests/ConnectivityTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Threading.Tasks;
 using NUnit.Framework;
 using Frends.SFTP.MoveFile.Enums;
 
@@ -9,26 +10,26 @@ namespace Frends.SFTP.MoveFile.Tests;
 public class ConnectivityTests : MoveFileTestBase
 {
     [Test]
-    public void MoveFile_TestWithPrivateKeyFileRsa()
+    public async Task MoveFile_TestWithPrivateKeyFileRsa()
     {
         _connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        _connection.PrivateKeyFilePassphrase = "passphrase";
+        _connection.PrivateKeyPassphrase = "passphrase";
         _connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
-        var result = SFTP.MoveFile(_input, _connection, default);
+        var result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsNotNull(result.Files);
     }
 
     [Test]
-    public void MoveFile_TestWithPrivateKeyFileRsaFromString()
+    public async Task MoveFile_TestWithPrivateKeyFileRsaFromString()
     {
         var key = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
 
         _connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-        _connection.PrivateKeyFilePassphrase = "passphrase";
+        _connection.PrivateKeyPassphrase = "passphrase";
         _connection.PrivateKeyString = key;
 
-        var result = SFTP.MoveFile(_input, _connection, default);
+        var result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsNotNull(result.Files);
     }
 }