+++ Frends.SFTP.MoveFile/Frends.SFTP.MoveFile.Tests/Lib/Helpers.cs
@@ -119,7 +119,7 @@ internal static class Helpers
 
     internal static string ToHex(byte[] bytes)
     {
-        StringBuilder result = new StringBuilder(bytes.Length * 2);
+        var result = new StringBuilder(bytes.Length * 2);
         for (int i = 0; i < bytes.Length; i++)
             result.Append(bytes[i].ToString("x2"));
         return result.ToString();