+++ Frends.SFTP.MoveFile/Frends.SFTP.MoveFile.Tests/MoveTests.cs
@@ -1,4 +1,5 @@
 using NUnit.Framework;
+using System.Threading.Tasks;
 using Frends.SFTP.MoveFile.Enums;
 
 namespace Frends.SFTP.MoveFile.Tests;
@@ -7,47 +8,47 @@ namespace Frends.SFTP.MoveFile.Tests;
 class MoveTests : MoveFileTestBase
 {
     [Test]
-    public void MoveFile_TestSimpleMove()
+    public async Task MoveFile_TestSimpleMove()
     {
-        var result = SFTP.MoveFile(_input, _connection, default);
+        var result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Files[0].DestinationPath));
     }
 
     [Test]
-    public void MoveFile_TestMoveOverwrite()
+    public async Task MoveFile_TestMoveOverwrite()
     {
-        var result = SFTP.MoveFile(_input, _connection, default);
+        var result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Files[0].DestinationPath));
         _input.IfTargetFileExists = FileExistsOperation.Overwrite;
 
         Helpers.GenerateDummyFile("test.txt");
-        result = SFTP.MoveFile(_input, _connection, default);
+        result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Files[0].DestinationPath));
     }
 
     [Test]
-    public void MoveFile_TestWithPatterns()
+    public async Task MoveFile_TestWithPatterns()
     {
         _input.Pattern = "*.txt";
-        var result = SFTP.MoveFile(_input, _connection, default);
+        var result = await SFTP.MoveFile(_input, _connection, default);
         Assert.AreEqual(1, result.Files.Count);
 
         Helpers.GenerateDummyFile("test.txt");
         _input.Pattern = "te*";
         _input.IfTargetFileExists = FileExistsOperation.Overwrite;
-        result = SFTP.MoveFile(_input, _connection, default);
+        result = await SFTP.MoveFile(_input, _connection, default);
         Assert.AreEqual(1, result.Files.Count);
     }
 
     [Test]
-    public void MoveFile_TestRename()
+    public async Task MoveFile_TestRename()
     {
-        var result = SFTP.MoveFile(_input, _connection, default);
+        var result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Files[0].DestinationPath));
         _input.IfTargetFileExists = FileExistsOperation.Rename;
 
         Helpers.GenerateDummyFile("test.txt");
-        result = SFTP.MoveFile(_input, _connection, default);
+        result = await SFTP.MoveFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Files[0].DestinationPath));
     }
 }