+++ Frends.SFTP.MoveFile/Frends.SFTP.MoveFile/Definitions/Input.cs
@@ -46,4 +46,29 @@ public class Input
     /// <example>true</example>
     [DefaultValue(true)]
     public bool CreateTargetDirectories { get; set; }
+
+    /// <summary>
+    /// Encoding for the read content.
+    /// By selecting 'Other' you can use any encoding.
+    /// </summary>
+    /// <example>FileEncoding.ANSI</example>
+    [DefaultValue(FileEncoding.ANSI)]
+    public FileEncoding FileEncoding { get; set; }
+
+    /// <summary>
+    /// Additional option for UTF-8 encoding to enable bom.
+    /// </summary>
+    /// <example>true</example>
+    [UIHint(nameof(FileEncoding), "", FileEncoding.UTF8)]
+    [DefaultValue(false)]
+    public bool EnableBom { get; set; }
+
+    /// <summary>
+    /// File encoding to be used.
+    /// Encoding don't support any unicode encoding. It only support the code page encodings. 
+    /// A partial list of possible encodings: https://en.wikipedia.org/wiki/Windows_code_page#List.
+    /// </summary>
+    /// <example>utf-8</example>
+    [UIHint(nameof(FileEncoding), "", FileEncoding.Other)]
+    public string EncodingInString { get; set; }
 }