+++ Frends.SFTP.MoveFile/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.1.0] - 2022-12-22
+### Updated
+- [Breaking ]Updated dependency SSH.NET to the newest version 2023.0.0.
+
+### Changed
+- Changed connection info builder to create the connection info as it's done in DownloadFiles.
+- [Breaking] Changed PrivateKeyFilePassphrase parameter to PrivateKeyPassphrase and enabled it when PrivateKeyString was used.
+
+### Added
+- Added FileEncoding for file names.
+
 ## [1.0.0] - 2023-22-05
 ### Added
 - Initial implementation